requirements.txt
discord.py
Pillow
google-cloud-firestore
flask
